Updated #witnesses 2nd paragraph
was:

After a Transmitter receives the challenge, it will beacon the challenge packet. There is no
intended recipient of this packet, and any Hotspot that is geographically proximate to the transmitter 
is eligible to witness it and report back to the blockchain. After a Hotspots witnesses a beacon from a 
transmitter, it will submit this as part of the POC Challenge receipt that gets assembled and submitted
to the blockchain by the POC Challenger.

Changed to:

After a Transmitter receives the challenge, it will beacon the challenge packet. There is no
intended recipient of this packet, and any Hotspot that is geographically proximate to the transmitter 
is eligible to witness it and report back to the blockchain. After a Hotspot witnesses a beacon from a 
transmitter, it will submit this as part of the POC Challenge receipt that gets assembled and submitted
to the blockchain by the POC Challenger.